requirements.txt
langchain-google-genai>=2.0.0
langgraph>=0.2.14
streamlit>=1.39.0
pydantic>=2.9.0
python-dotenv>=1.0.1
pandas>=2.2.0
